+++ app/controllers/um/organization.controller.js
@@ -3,41 +3,56 @@ const Organization = require('../../models/um/organization.model.js');
 
 // Create and Save a new Organization
 exports.create = (req, res) => {
-    // Validate request
-    
-    if(!req.body) {
-        return res.status(400).send({
-            message: "Organization body can not be empty"
-        });
+    if (req.session.user && req.cookies.user_sid) {
+        // Validate request
+
+        if (!req.body) {
+            return res.status(400).send({
+                message: "Organization body can not be empty"
+            });
+        }
+
+        // Create a Organization
+        const organization = new Organization(req.body);
+        req.body['org_status'] = "Active";
+        // Save Organization in the database
+        organization.save()
+            .then(data => {
+                res.send(data);
+            }).catch(err => {
+                res.status(500).send({
+                    message: err.message || "Some error occurred while creating the Organization."
+                });
+            });
+    } else {
+        res.redirect('/api/login');
     }
 
-    // Create a Organization
-    const organization = new Organization(req.body);
-    req.body['org_status'] = "Active";
-    // Save Organization in the database
-    organization.save()
-    .then(data => {
-        res.send(data);
-    }).catch(err => {
-        res.status(500).send({
-            message: err.message || "Some error occurred while creating the Organization."
-        });
-    });
+
+
+
 };
 
 
 
 // Retrieve and return all organizations from the database.
 exports.findAll = (req, res) => {
-    Organization.find()
-    .then(organizations => {
-        console.log(organizations)
-        res.send(organizations);
-    }).catch(err => {
-        res.status(500).send({
-            message: err.message || "Some error occurred while retrieving organizations."
-        });
-    });
+    if (req.session.user && req.cookies.user_sid) {
+        Organization.find()
+            .then(organizations => {
+                res.send(organizations);
+            }).catch(err => {
+                res.status(500).send({
+                    message: err.message || "Some error occurred while retrieving organizations."
+                });
+            });
+    } else {
+        res.redirect('/api/login');
+    }
+
+
+
+
 };
 
 
@@ -46,25 +61,33 @@ exports.findAll = (req, res) => {
 
 // Find a single organization with a id
 exports.findOne = (req, res) => {
-    console.log(req.params.id)
-    Organization.findById(req.params.id)
-    .then(organization => {
-        if(!organization) {
-            return res.status(404).send({
-                message: "Organization not found " + req.params.id
-            });            
-        }
-        res.send(organization);
-    }).catch(err => {
-        if(err.kind === 'ObjectId') {
-            return res.status(404).send({
-                message: "Organization not found " + req.params.id
-            });                
-        }
-        return res.status(500).send({
-            message: "Error retrieving Organization " + req.params.id
-        });
-    });
+    if (req.session.user && req.cookies.user_sid) {
+        
+        Organization.findById(req.params.id)
+            .then(organization => {
+                if (!organization) {
+                    return res.status(404).send({
+                        message: "Organization not found " + req.params.id
+                    });
+                }
+                res.send(organization);
+            }).catch(err => {
+                if (err.kind === 'ObjectId') {
+                    return res.status(404).send({
+                        message: "Organization not found " + req.params.id
+                    });
+                }
+                return res.status(500).send({
+                    message: "Error retrieving Organization " + req.params.id
+                });
+            });
+    } else {
+        res.redirect('/api/login');
+    }
+
+
+
+
 };
 
 
@@ -73,34 +96,43 @@ exports.findOne = (req, res) => {
 
 // Update a Organization identified by the id in the request
 exports.update = (req, res) => {
-    // Validate Request
-    if(!req.body) {
-        return res.status(400).send({
-            message: "Organization body can not be empty"
-        });
-    }
-
-    // Find Organization and update it with the request body
-    console.log(req.body)
-    Organization.findByIdAndUpdate(req.params.id, req.body , {new: true})
-    .then(organization => {
-        if(!organization) {
-            return res.status(404).send({
-                message: "Organization not found with id " + req.params.id
+    if (req.session.user && req.cookies.user_sid) {
+        // Validate Request
+        if (!req.body) {
+            return res.status(400).send({
+                message: "Organization body can not be empty"
             });
         }
-        console.log(organization)
-        res.send(organization);
-    }).catch(err => {
-        if(err.kind === 'ObjectId') {
-            return res.status(404).send({
-                message: "Organization not found with id " + req.params.id
-            });                
-        }
-        return res.status(500).send({
-            message: "Error updating Organization with id " + req.params.id
-        });
-    });
+
+        // Find Organization and update it with the request body
+        
+        Organization.findByIdAndUpdate(req.params.id, req.body, { new: true })
+            .then(organization => {
+                if (!organization) {
+                    return res.status(404).send({
+                        message: "Organization not found with id " + req.params.id
+                    });
+                }
+                
+                res.send(organization);
+            }).catch(err => {
+                if (err.kind === 'ObjectId') {
+                    return res.status(404).send({
+                        message: "Organization not found with id " + req.params.id
+                    });
+                }
+                return res.status(500).send({
+                    message: "Error updating Organization with id " + req.params.id
+                });
+            });
+    } else {
+        res.redirect('/api/login');
+    }
+
+
+
+
+
 };
 
 
@@ -109,25 +141,33 @@ exports.update = (req, res) => {
 
 // Delete a organization with the specified organizationId in the request
 exports.delete = (req, res) => {
-    console.log(req.body)
-    Organization.findByIdAndRemove(req.params.id)
-    .then(organization => {
-        if(!organization) {
-            return res.status(404).send({
-                message: "Organization not found with id " + req.params.id
+    if (req.session.user && req.cookies.user_sid) {
+        
+        Organization.findByIdAndRemove(req.params.id)
+            .then(organization => {
+                if (!organization) {
+                    return res.status(404).send({
+                        message: "Organization not found with id " + req.params.id
+                    });
+                }
+                
+                res.send({ message: "Organization deleted successfully!" });
+            }).catch(err => {
+                if (err.kind === 'ObjectId' || err.name === 'NotFound') {
+                    return res.status(404).send({
+                        message: "Organization not found with id " + req.params.id
+                    });
+                }
+                
+                return res.status(500).send({
+                    message: "Could not delete organization with id " + req.params.id
+                });
             });
-        }
-        console.log(organization)
-        res.send({message: "Organization deleted successfully!"});
-    }).catch(err => {
-        if(err.kind === 'ObjectId' || err.name === 'NotFound') {
-            return res.status(404).send({
-                message: "Organization not found with id " + req.params.id
-            });                
-        }
-        console.log(organization)
-        return res.status(500).send({
-            message: "Could not delete organization with id " + req.params.id
-        });
-    });
+    } else {
+        res.redirect('/api/login');
+    }
+
+
+
+
 };
